Fix segfault related to round robin  pop() and push()
In pop(), the check for the sentinel node was written incorrectly. This caused a bug where the sentinel->next node would
never get set to NULL when the last task node was popped off, causing a segfault.

In push(), after traversing the queue, the new task node was inserted at the end of the queue. However, it's next pointer was not
being properly updated to point back to the sentinel (completing the circular link).