Fixed task timings in rr.c
In rr.c:round_robin(), runtime update should run after taskRunTime is calculated. This was causing the task
timings to be incorrect. This is now corrected.

In rrtests.c, a longer unit-test (customRR2) was written to test the round robin queue with more tasks.
The unit-test asserts that the waiting_time and turnaround_time for each task is correct.